FastAPI/requirements.txt
fastapi==0.75.1
uvicorn[standard]==0.18.2
sqlalchemy==1.4.31
psycopg2-binary==2.9.3
matplotlib==3.5.1
numpy<2
jinja2>=3.0.0
passlib
python-multipart
python-jose[cryptography]